Rename file and add e2e tests for staff selected doc types by team name endpoint
